Fix incorrect argument type

+++ local_planner/include/local_planner/lattice.h
@@ -86,7 +86,7 @@ public:
 
     std::vector<int> getReferenceWaypointIds(const int nearest_wp_id) const;
 
-    int getWaypointIdAtDistance(const int start_id, const int target_dist) const;
+    int getWaypointIdAtDistance(const int start_id, const double target_dist) const;
 
     std::vector<std::vector<Vertex>> generateLayers(const std::vector<int>& ref_waypoint_ids, const double source_x,
                                                     const double source_y) const;
